+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## master
 
-* Added ability to replay asciicast from stdin by passing "-" as filename
+* Playing asciicast from stdin: `cat demo.json | asciinema play -`
+* Playing asciicast from IPFS: `asciinema play ipfs://QmcdXYJp6e4zNuimuGeWPwNMHQdxuqWmKx7NhZofQ1nw2V`
 * `-q/--quiet` option added to `rec` command
 
 ## 1.1.1 (2015-06-21)